+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# Unreleased
+
+### Added
+
+- Typst packages support (unstable) ([#51])
+- `with-typst-packages-unpublished` template ([#51])
+
+### Changed
+
+- Use `unstableTypstPackages` in `with-typst-packages` template ([#51])
+
+[#51]: https://github.com/loqusion/typix/pull/51
+
 # [0.3.0] (January 27, 2025)
 
 ### Added